fix: use spot price when calculating principal before unwrap
we were using oracle price and that is not inline with the actual decrease liquidity operation on the pool that uses the spot price. The discrenpency between oracle price and pool spot price can be made as big as the attacker wants to gulp all of the fees for themselevs if they are not the sole owner of Liquidity Position. For normal users, because we expect that the difference between amountReceived and amount calculated principal is fees, it can underflow and the operation can fail. That's what was happening for the testcase testUnwrap_Unichain.