fix: use current tokenId balance as an indication weather to skip it
Earlier, we were using totalSupply and in balance of and that is a weak condition. A tokenId can be wrapped by someone else not held by the user that has enabled it as collateral. In this case, it should not compute the value of tokenId because the balance is anyway zero. Additionally, we added the same skipping logic transfer as well to avoid zero transfers