WIP: add a way to test fees related logic
Using flashloan, in v3pools, we can donate fees to active liquidity providers simillar to the donate function in PoolManager. We are now using that in invaraint tests to make sure nothing breaks. Right now, the tests are not passing likely because calculateExactedValueOfTokenIdAfterUnwrap in MockUniswapV4Wrapper likley overestimates the value of the collateral after the unwrap is done and the partial unwrap fails without catching it