test: workaround to make sure calculateExactedValueOfTokenIdAfterUnwrap works the same as the actual unwrap
we unwrap 0 amount for v4 wrapper before the actual unwrap. unwrapping the 0 amount will convert the pending fees into feesOwed tracked in the v4 wrapper contract. this way the calculation in calculateExactedValueOfTokenIdAfterUnwrap is the exact same the actual unwrap call. if not done then there will be an error of 1 wei and that is problematic because then we can't accurately predict if the next call will fails or not