feat: use previewUnwrap in calculateValueOfTokenId
what we were doing before is that we were calculating the amount0 and amount 1 for the entire tokenId and then after coverting them into usd amount and adding them up, we were taking the proportion. This has less rounding errors and that is good for the user but in the actual unwrap, that is not what happens. In the actual unwrap, we take the propotion of the liquidity first and decrease that much liquidity and that't the principal user gets back. We then calculate the pending fees for the whole tokenId and send the proportional fees to the user. If the spot price and oracle price is the same, we want to make sure the balanceOf correcty reflects that actual tokens gotten from actual unwrap. If user is ever able to inflate the value of one unit of liquidity we want to make sure that balanceOf calculates the value of tokenId in the same way